refactor: 홈 관련 컴포넌트 폴더 이동, 일부 공통 컴포넌트로 분리
