+++ app/[locale]/(main)/(home)/components/MentorCardList.tsx
@@ -1,9 +1,9 @@
 import { useGetMentorList } from "@/apis/user/hooks/useGetMentorList";
 import { GetMentorListRequest } from "@/apis/user/types";
+import { UserCard } from "@/components/UserCard";
 import { PATH } from "@/constants/path";
 import { useIntersect } from "@/hooks/useIntersect";
 import { Link } from "@/libs/navigation";
-import { UserCard } from "./UserCard";
 
 interface MentorCardListProps {
   params: Pick<GetMentorListRequest, "languages">;
