+++ app/[locale]/(main)/(home)/components/ProfileCompleteBanner.tsx
