+++ app/[locale]/(main)/(home)/components/SwitchExploreUserButton.tsx
