+++ app/[locale]/(main)/coffeechat/[id]/MentorView.tsx
@@ -3,8 +3,8 @@ import useGetCoffeeChatById from "@/apis/coffeechat/hooks/useGetCoffeeChatById";
 import useCancelCoffeeChat from "@/app/[locale]/(main)/coffeechat/hooks/useCancelCoffeeChat";
 import { BottomSheet } from "@/components/BottomSheet";
 import { Button, LinkButton } from "@/components/Button";
+import { MenteeProfile } from "@/components/UserProfile";
 import { PATH } from "@/constants/path";
-import { MenteeProfile } from "../../components/UserProfile";
 import { CoffeeChatSchedule } from "../components/CoffeeChatSchedule";
 import { CoffeeChatTypeSelect } from "../components/CoffeeChatTypeSelectBottomSheet";
 import { QuestionToMentee } from "../components/QuestionToMenteeBottomSheet/QuestionToMenteeBottomSheet";
