+++ app/[locale]/(main)/profile/mentor/[id]/page.tsx
@@ -2,9 +2,9 @@
 
 import { useTranslations } from "next-intl";
 import { useGetMentorById } from "@/apis/user/hooks/useGetMentorById";
-import { MentorProfile } from "@/app/[locale]/(main)/components/UserProfile";
 import { GoToLoginBottomSheet } from "@/app/components/GoToLoginBottomSheet";
 import { LinkButton } from "@/components/Button";
+import { MentorProfile } from "@/components/UserProfile";
 import { PATH } from "@/constants/path";
 import { useAuth } from "@/hooks/useAuth";
 import { useToggle } from "@/hooks/useToggle";
